Logging (task 13)

See merge request VasyaDen/fox-university!7

+++ src/main/java/com/foxminded/foxuniversity/service/implementation/GroupServiceImpl.java
@@ -6,6 +6,8 @@ import com.foxminded.foxuniversity.domain.Group;
 import com.foxminded.foxuniversity.domain.Lesson;
 import com.foxminded.foxuniversity.service.GroupService;
 import com.foxminded.foxuniversity.service.StudentService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -18,8 +20,11 @@ public class GroupServiceImpl implements GroupService {
     @Autowired
     private StudentService studentService;
 
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
     @Override
     public List<Group> getAll() {
+        logger.debug("GroupService calls groupDao.getAll().");
         List<Group> groups = groupDao.getAll();
         fillStudents(groups);
         return groups;
@@ -27,6 +32,7 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public Group getById(int id) {
+        logger.debug("GroupService calls groupDao.getById(" + id + ").");
         Group group = groupDao.getById(id);
         fillStudents(group);
         return group;
@@ -34,6 +40,7 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public List<Group> getByLesson(Lesson lesson) {
+        logger.debug("GroupService calls groupDao.getByLesson(Lesson{id = " + lesson.getId() + "}).");
         List<Group> groups = groupDao.getByLesson(lesson);
         fillStudents(groups);
         return groups;
@@ -41,6 +48,7 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public List<Group> getByCourse(Course course) {
+        logger.debug("GroupService calls groupDao.getByCourse(Course{id = " + course.getId() + "}).");
         List<Group> groups = groupDao.getByCourse(course);
         fillStudents(groups);
         return groups;
@@ -48,51 +56,75 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public void save(Group group) {
+        logger.debug("GroupService calls groupDao.save(" + group + ").");
         groupDao.save(group);
     }
 
     @Override
     public boolean update(Group group) {
+        logger.debug("GroupService calls groupDao.update(" + group + ").");
         return groupDao.update(group);
     }
 
     @Override
     public boolean delete(Group group) {
+        logger.debug("GroupService calls groupDao.delete(" + group + ").");
         return groupDao.delete(group);
     }
 
     @Override
     public boolean assignToCourses(Group group, List<Course> courses) {
+        logger.debug("GroupService calls groupDao.assignToCourses(" + group + ", List<Course>[" + courses.size() + "])");
         if (groupDao.assignToCourses(group, courses)) {
+            logger.debug("Assignment was successful. Set group to each course in the list.");
             for (Course course : courses) {
                 course.getGroups().add(group);
             }
             return true;
         }
+        logger.warn("Assignment " + group + " to courses was cancelled.");
         return false;
     }
 
     @Override
     public boolean deleteFromCourse(Group group, Course course) {
-        if (groupDao.deleteFromCourse(group, course)) {
-            course.getGroups().remove(group);
-            return true;
+        if (course.getGroups().contains(group)) {
+            logger.debug("Call groupDao.deleteFromCourse(" + group + ", Course{id = " + course.getId() + "}).");
+            if (groupDao.deleteFromCourse(group, course)) {
+                logger.debug("Deletion was successful. Remove group from course.");
+                course.getGroups().remove(group);
+                return true;
+            }
+            logger.warn("Deletion " + group + " from Course{id = " + course.getId() + " was cancelled in DAO-layer.");
+            return false;
         }
+        logger.warn("Deletion was cancelled: " + group + " isn't assigned to the Course{id = " + course.getId() + "}.");
         return false;
     }
 
     @Override
     public boolean deleteFromCourse(Group group, List<Course> courses) {
+        for (Course course : courses) {
+            if (!course.getGroups().contains(group)) {
+                logger.warn("Deletion was cancelled: " + group +
+                        " isn't assigned to the Course{id = " + course.getId() + "}.");
+                return false;
+            }
+        }
+        logger.debug("Call groupDao.deleteFromCourse(" + group + ", Courses[" + courses.size() + "]).");
         if (groupDao.deleteFromCourse(group, courses)) {
+            logger.debug("Deletion was successful. Remove group from courses.");
             for (Course course : courses) {
                 course.getGroups().remove(group);
             }
             return true;
         }
+        logger.warn("Deletion " + group + " from courses was cancelled in DAO-layer.");
         return false;
     }
 
     private void fillStudents(Group group) {
+        logger.debug("Set students to group: call studentService.getByGroup(" + group + ")");
         group.setStudents(studentService.getByGroup(group));
     }
